make it compatible for floating values

+++ script.js
@@ -4,7 +4,7 @@ const result = document.querySelector('#result')
 form.addEventListener('submit',(e)=>{
     e.preventDefault();
 
-    const temp = parseInt(document.querySelector('#temp').value);
+    const temp = parseFloat(document.querySelector('#temp').value);
     if(document.querySelector('#toFahrenhiet').checked){
         const convertedTemp = ((temp*(9/5)) + 32).toFixed(2);
         result.innerHTML = `Your result is ${convertedTemp} °F`;
